requirements.txt
# AI Agent Project - Root Dependencies
# This file consolidates all service dependencies for development environment
# Individual services have their own requirements.txt for production builds

# ===========================================
# Common Dependencies
# ===========================================
kafka-python>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# ===========================================
# API Server
# ===========================================
fastapi>=0.100.0
uvicorn>=0.20.0

# ===========================================
# Search Worker
# ===========================================
trafilatura>=1.0.0
duckduckgo-search>=4.0.0

# ===========================================
# AI Worker (GPU required)
# ===========================================
# Note: vllm and torch are heavy dependencies, 
# install separately on GPU machines only
# vllm==0.6.3
# torch>=2.0.0

# ===========================================
# Development Tools
# ===========================================
pytest>=7.0.0
black>=23.0.0
isort>=5.0.0
httpx>=0.24.0  # For FastAPI TestClient
